+++ app/models/service_to_category.rb
@@ -1,0 +1,4 @@
+class ServiceToCategory < ApplicationRecord
+  belongs_to :service
+  belongs_to :service_category
+end